Tag detail tests

+++ backend/models.py
@@ -7,6 +7,7 @@ from django.dispatch import receiver
 
 from azure.storage.blob import BlockBlobService
 
+
 class Address(gis_models.Model):
     objects = gis_models.GeoManager()
     point = gis_models.PointField(srid=4326)
@@ -34,6 +35,7 @@ class Address(gis_models.Model):
     def longitude(self):
         return self.point.x
 
+
 class Tag(base_models.Model):
     address = base_models.ForeignKey(Address, on_delete=base_models.CASCADE)
     creator_user_id = base_models.CharField(max_length=255)
@@ -51,8 +53,12 @@ class Tag(base_models.Model):
     tag_words = base_models.CharField(max_length=255, blank=True)
     tag_initials = base_models.CharField(max_length=20, blank=True)
 
+
 @receiver(pre_delete, sender=Tag)
 def delete_image(sender, instance, **kwargs):
-    image_name = instance.img.split("/")[-1]
-    block_blob_service = BlockBlobService(account_name=os.environ['AZURE_IMAGE_CONTAINER_NAME'], account_key=os.environ['AZURE_IMAGE_CONTAINER_KEY'])
-    block_blob_service.delete_blob('images', image_name)
+    # TODO: figure out why `settings.DEBUG` appears false despite this logic returning `True`
+    debug = True if os.environ.get("DEBUG") else False
+    if not debug:
+        image_name = instance.img.split("/")[-1]
+        block_blob_service = BlockBlobService(account_name=os.environ['AZURE_IMAGE_CONTAINER_NAME'], account_key=os.environ['AZURE_IMAGE_CONTAINER_KEY'])
+        block_blob_service.delete_blob('images', image_name)
